# week10/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"email": {
						"type": "string",
						"description": "유저 이메일"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					},
					"preferCategory": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "선호 카테고리"
					}
				},
				"required": [
					"email",
					"name",
					"preferCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserSignUpResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserSignUpResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string",
						"description": "유저 이메일",
						"example": "test@example.com"
					},
					"name": {
						"type": "string",
						"description": "유저 이름",
						"example": "홍길동"
					},
					"gender": {
						"type": "string",
						"description": "성별",
						"example": "MALE"
					},
					"birth": {
						"type": "string",
						"description": "생년월일",
						"example": "2000-01-01"
					},
					"address": {
						"type": "string",
						"description": "주소 (선택)",
						"example": "서울시 강남구"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세주소 (선택)",
						"example": "역삼동 123-45"
					},
					"password": {
						"type": "string",
						"description": "비밀번호",
						"example": "password123"
					},
					"phoneNumber": {
						"type": "string",
						"description": "전화번호",
						"example": "010-1234-5678"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "선호 카테고리 ID 배열",
						"example": [
							1,
							2
						]
					}
				},
				"required": [
					"email",
					"name",
					"gender",
					"birth",
					"password",
					"phoneNumber",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"email": {
						"type": "string",
						"description": "유저 이메일"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					}
				},
				"required": [
					"id",
					"email",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserUpdateResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserUpdateResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserUpdateResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateRequest": {
				"properties": {
					"phoneNumber": {
						"type": "string",
						"description": "전화번호 (선택)",
						"example": "010-1234-5678"
					},
					"birth": {
						"type": "string",
						"description": "생년월일 (선택)",
						"example": "2000-01-01"
					},
					"address": {
						"type": "string",
						"description": "주소 (선택)",
						"example": "서울시 강남구"
					},
					"gender": {
						"type": "string",
						"description": "성별 (선택)",
						"example": "MALE"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "유저 미션 ID"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"status": {
						"type": "string",
						"description": "미션 상태 (CHALLENGING / DONE)"
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionListResponse": {
				"properties": {
					"userMissions": {
						"items": {
							"$ref": "#/components/schemas/UserMissionAddResponse"
						},
						"type": "array",
						"description": "유저 미션 목록"
					}
				},
				"required": [
					"userMissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionListResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionListApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserMissionCompleteResponse": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double",
						"description": "업데이트된 미션 수"
					}
				},
				"required": [
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionCompleteResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionCompleteResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionCompleteApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionCompleteResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playground-umc-10th-nodejs",
		"version": "1.0.0",
		"description": "UMC 테스트 프로젝트",
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "회원가입 성공",
											"data": {
												"email": "test@example.com",
												"name": "홍길동",
												"preferCategory": [
													"1",
													"2"
												]
											}
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "이미 존재하는 이메일입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserSignUpApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 409,
											"message": "이미 존재하는 이메일입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserSignUpApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "회원가입 API",
				"summary": "회원가입을 처리하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
		"/users/me": {
			"patch": {
				"operationId": "HandleUserUpdate",
				"responses": {
					"200": {
						"description": "사용자 정보 수정 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserUpdateResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "사용자 정보 수정 성공",
											"data": {
												"id": 1,
												"email": "test@example.com",
												"name": "홍길동"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요합니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserUpdateApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 401,
											"message": "로그인이 필요합니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 사용자입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserUpdateApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 사용자입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserUpdateApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "사용자 정보 수정 API",
				"summary": "로그인한 사용자의 정보를 수정하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/users/missions": {
			"get": {
				"operationId": "HandleGetUserMissions",
				"responses": {
					"200": {
						"description": "미션 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionListResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "미션 목록 조회 성공",
											"data": {
												"userMissions": [
													{
														"id": 1,
														"userId": 1,
														"missionId": 1,
														"status": "CHALLENGING"
													}
												]
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요합니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 401,
											"message": "로그인이 필요합니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "진행 중인 미션 목록 조회 API",
				"summary": "로그인한 사용자의 진행 중인 미션 목록을 조회하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "페이지네이션 커서",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/missions/{missionId}": {
			"patch": {
				"operationId": "HandleCompleteUserMission",
				"responses": {
					"200": {
						"description": "미션 완료 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionCompleteResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "미션 완료 성공",
											"data": {
												"count": 1
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "진행 중인 미션이 아닙니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionCompleteApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 400,
											"message": "진행 중인 미션이 아닙니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요합니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionCompleteApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 401,
											"message": "로그인이 필요합니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 사용자 또는 미션입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionCompleteApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 사용자 또는 미션입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionCompleteApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 완료 API",
				"summary": "진행 중인 미션을 완료로 변경하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}