# week10/src/modules/users/controllers/user.controller.ts
import { Body, Controller, Get, Patch, Path, Post, Put, Queries, Route, Tags, Middlewares, Request, Response, SuccessResponse, Example } from "tsoa";
import {
  UserMissionListResponse,
  UserMissionCompleteResponse,
  UserMissionListApiResponse,
  UserMissionCompleteApiResponse,
  UserSignUpRequest,
  UserSignUpResponse,
  UserSignUpApiResponse,
  UserUpdateRequest,
  UserUpdateResponse,
  UserUpdateApiResponse,
} from "../dtos/user.dto.js";
import { userSignUp, userUpdate, userMissionList, userMissionComplete } from "../services/user.service.js";
import { ApiResponse, success } from "../../../common/responses/response.js";
import { StatusCodes } from "http-status-codes";
import { authorizeUser, authenticateJwt } from "../../../common/middlewares/auth.middleware.js";
import { Request as ExpressRequest } from "express";
import { AppError } from "../../../common/errors/app.error.js";

interface UserMissionListQuery {
  /** 페이지네이션 커서 */
  cursor?: number;
}

@Route("users")
@Tags("Users")
export class UserController extends Controller {
  /**
   * 회원가입 API
   * @summary 회원가입을 처리하는 엔드포인트입니다.
   */
  @Post("signup")
  @SuccessResponse(200, "회원가입 성공")
  @Example<UserSignUpApiResponse>({
    success: true,
    statusCode: 200,
    message: "회원가입 성공",
    data: {
      email: "test@example.com",
      name: "홍길동",
      preferCategory: ["1", "2"],
    },
  })
  @Response<UserSignUpApiResponse>(409, "이미 존재하는 이메일입니다.", {
    success: false,
    statusCode: 409,
    message: "이미 존재하는 이메일입니다.",
    data: null,
  })
  @Response<UserSignUpApiResponse>(500, "서버 내부 오류", {
    success: false,
    statusCode: 500,
    message: "서버 내부 오류",
    data: null,
  })
  public async handleUserSignUp(
    @Body() body: UserSignUpRequest,
  ): Promise<ApiResponse<UserSignUpResponse>> {
    try {
      const user = await userSignUp(body);
      return success(StatusCodes.OK, "회원가입 성공", user);
    } catch (err) {
      if (err instanceof AppError) {
        throw err;
      }
      throw new AppError(
        StatusCodes.INTERNAL_SERVER_ERROR,
        (err as Error).message || "서버 내부 오류"
      );
    }
  }

  /**
   * 사용자 정보 수정 API
   * @summary 로그인한 사용자의 정보를 수정하는 엔드포인트입니다.
   */
  @Patch("me")
  @Middlewares(authenticateJwt())
  @SuccessResponse(200, "사용자 정보 수정 성공")
  @Example<UserUpdateApiResponse>({
    success: true,
    statusCode: 200,
    message: "사용자 정보 수정 성공",
    data: { id: 1, email: "test@example.com", name: "홍길동" },
  })
  @Response<UserUpdateApiResponse>(401, "로그인이 필요합니다.", {
    success: false,
    statusCode: 401,
    message: "로그인이 필요합니다.",
    data: null,
  })
  @Response<UserUpdateApiResponse>(404, "존재하지 않는 사용자입니다.", {
    success: false,
    statusCode: 404,
    message: "존재하지 않는 사용자입니다.",
    data: null,
  })
  @Response<UserUpdateApiResponse>(500, "서버 내부 오류", {
    success: false,
    statusCode: 500,
    message: "서버 내부 오류",
    data: null,
  })
  public async handleUserUpdate(
    @Body() body: UserUpdateRequest,
    @Request() req: ExpressRequest,
  ): Promise<ApiResponse<UserUpdateResponse>> {
    try {
      const userId = Number((req.user as any).id);
      const user = await userUpdate(userId, body);
      return success(StatusCodes.OK, "사용자 정보 수정 성공", user);
    } catch (err) {
      if (err instanceof AppError) {
        throw err;
      }
      throw new AppError(
        StatusCodes.INTERNAL_SERVER_ERROR,
        (err as Error).message || "서버 내부 오류"
      );
    }
  }

  /**
   * 진행 중인 미션 목록 조회 API
   * @summary 로그인한 사용자의 진행 중인 미션 목록을 조회하는 엔드포인트입니다.
   */
  @Get("missions")
  @Middlewares(authenticateJwt())
  @SuccessResponse(200, "미션 목록 조회 성공")
  @Example<UserMissionListApiResponse>({
    success: true,
    statusCode: 200,
    message: "미션 목록 조회 성공",
    data: { userMissions: [{ id: 1, userId: 1, missionId: 1, status: "CHALLENGING" }] },
  })
  @Response<UserMissionListApiResponse>(401, "로그인이 필요합니다.", {
    success: false,
    statusCode: 401,
    message: "로그인이 필요합니다.",
    data: null,
  })
  @Response<UserMissionListApiResponse>(500, "서버 내부 오류", {
    success: false,
    statusCode: 500,
    message: "서버 내부 오류",
    data: null,
  })
  public async handleGetUserMissions(
    @Request() req: ExpressRequest,
    @Queries() query: UserMissionListQuery,
  ): Promise<ApiResponse<UserMissionListResponse>> {
    try {
      const userId = Number((req.user as any).id);
      const userMissions = await userMissionList(userId, query.cursor);
      return success(StatusCodes.OK, "미션 목록 조회 성공", userMissions);
    } catch (err) {
      if (err instanceof AppError) {
        throw err;
      }
      throw new AppError(
        StatusCodes.INTERNAL_SERVER_ERROR,
        (err as Error).message || "서버 내부 오류"
      );
    }
  }

  /**
   * 미션 완료 API
   * @summary 진행 중인 미션을 완료로 변경하는 엔드포인트입니다.
   */
  @Patch("missions/{missionId}")
  @Middlewares(authenticateJwt())
  @SuccessResponse(200, "미션 완료 성공")
  @Example<UserMissionCompleteApiResponse>({
    success: true,
    statusCode: 200,
    message: "미션 완료 성공",
    data: { count: 1 },
  })
  @Response<UserMissionCompleteApiResponse>(401, "로그인이 필요합니다.", {
    success: false,
    statusCode: 401,
    message: "로그인이 필요합니다.",
    data: null,
  })
  @Response<UserMissionCompleteApiResponse>(404, "존재하지 않는 사용자 또는 미션입니다.", {
    success: false,
    statusCode: 404,
    message: "존재하지 않는 사용자 또는 미션입니다.",
    data: null,
  })
  @Response<UserMissionCompleteApiResponse>(400, "진행 중인 미션이 아닙니다.", {
    success: false,
    statusCode: 400,
    message: "진행 중인 미션이 아닙니다.",
    data: null,
  })
  @Response<UserMissionCompleteApiResponse>(500, "서버 내부 오류", {
    success: false,
    statusCode: 500,
    message: "서버 내부 오류",
    data: null,
  })
  public async handleCompleteUserMission(
    @Path() missionId: number,
    @Request() req: ExpressRequest,
  ): Promise<ApiResponse<UserMissionCompleteResponse>> {
    try {
      const userId = Number((req.user as any).id);
      const result = await userMissionComplete(userId, missionId);
      return success(StatusCodes.OK, "미션 완료 성공", result);
    } catch (err) {
      if (err instanceof AppError) {
        throw err;
      }
      throw new AppError(
        StatusCodes.INTERNAL_SERVER_ERROR,
        (err as Error).message || "서버 내부 오류"
      );
    }
  }

  @Get("guest")
  public async handleGuestPage(): Promise<String> {
    return `
      <h1>게스트 페이지</h1>
      <p>이 페이지는 로그인이 필요 없습니다.</p>
      <ul>
        <li><a href="/api/v1/users/mypage">마이페이지 (로그인 필요)</a></li>
      </ul>
    `;
  }

  @Get("login")
  public async handleLoginPage(): Promise<String> {
    return "<h1>로그인 페이지</h1><p>로그인이 필요한 페이지에서 튕겨나오면 여기로 옵니다.</p>";
  }

  @Get("mypage")
  @Middlewares(authorizeUser())
  public async handleMypage(@Request() req: ExpressRequest): Promise<String> {
    return `
      <h1>마이페이지</h1>
      <p>환영합니다, ${req.cookies.username}님!</p>
      <p>이 페이지는 로그인한 사람만 볼 수 있습니다.</p>
    `;
  }

  @Get("set-login")
  public async handleSetLogin(@Request() req: ExpressRequest): Promise<String> {
    req.res!.cookie("username", "UMC10th", { maxAge: 3600000 });
    return '로그인 쿠키(username=UMC10th) 생성 완료! <a href="/api/v1/users/mypage">마이페이지로 이동</a>';
  }

  @Get("set-logout")
  public async handleSetLogout(@Request() req: ExpressRequest): Promise<String> {
    req.res!.clearCookie("username");
    return '로그아웃 완료 (쿠키 삭제). <a href="/api/v1/users/guest">메인으로</a>';
  }
}